+++ plugins/plugin-chart-pie/src/Pie.tsx
@@ -1,19 +1,35 @@
-import React, { useState, FC, useEffect, memo, useRef } from 'react';
-import { styled } from '@superset-ui/style';
-import { t } from '@superset-ui/translation';
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+import React, { useState, FC, useEffect, memo, useRef, useCallback } from 'react';
+import { styled, t, CategoricalColorNamespace } from '@superset-ui/core';
 import {
   PieChart,
-  ResponsiveContainer,
   Pie as RechartsPie,
   Cell,
   RechartsFunction,
-  PieLabelRenderProps,
   Legend,
   PieProps as RechartsPieProps,
   LegendType,
 } from 'recharts';
-import { CategoricalColorNamespace } from '@superset-ui/color';
-import { LegendPosition, renderActiveShape, getLegendProps, LABELS_MARGIN } from './utils';
+import { LegendPosition, renderActiveShape, getLegendProps, ActiveShapeProps } from './utils';
+
+export const LABELS_MARGIN = 100;
 
 type EventData = {
   color: string;
@@ -88,42 +104,48 @@ const Pie: FC<PieProps<string, string>> = memo(props => {
   } = props;
   const [notification, setNotification] = useState<string | null>(null);
   const [activeIndex, setActiveIndex] = useState<number>(0);
-  const [isAnimationEnded, setIsAnimationEnded] = useState<boolean>(false);
-  const [exploreCounter, setExploreCounter] = useState<number>(0);
   const [disabledDataKeys, setDisabledDataKeys] = useState<string[]>([]);
   const [legendWidth, setLegendWidth] = useState<number | null>(0);
+  const [updater, setUpdater] = useState<number>(0);
   const rootRef = useRef<HTMLDivElement>(null);
   const isSideLegend =
     showLegend && (legendPosition === LegendPosition.right || legendPosition === LegendPosition.left);
 
   const currentData = data.filter(item => !disabledDataKeys.includes(item[groupBy]));
-
-  useEffect(() => {
-    if (isAnimationEnded) {
-      setExploreCounter(exploreCounter + 1);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [showLabels, showLegend, isDonut, legendPosition]);
+  const isExplore = window.location.pathname.includes('/explore');
 
   useEffect(() => {
     if (isSideLegend && rootRef.current) {
       const legend = rootRef.current.querySelector('.recharts-legend-wrapper');
-      setLegendWidth(legend?.getBoundingClientRect()?.width || null);
+      const currentWidth = legend?.getBoundingClientRect()?.width || null;
+      if (currentWidth !== legendWidth) {
+        setLegendWidth(currentWidth ? currentWidth + 20 : currentWidth);
+      }
     }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [rootRef.current]);
+  }, [updater, isSideLegend, legendWidth]);
+
+  useEffect(() => {
+    if (isSideLegend) {
+      setLegendWidth(0);
+    }
+  }, [isSideLegend, props]);
+
+  const forceUpdate = useCallback(() => setUpdater(Math.random()), []);
+
+  useEffect(() => {
+    forceUpdate();
+  }, [forceUpdate, props]);
 
   const onPieEnter = (data: object, index: number) => setActiveIndex(index);
-
   const closeNotification = () => setNotification(null);
-  const { getColor } = CategoricalColorNamespace;
 
   const onClick = () => {
-    // eslint-disable-next-line no-restricted-globals
-    if (location.pathname.includes('/explore')) {
+    if (isExplore) {
       setNotification(t('Sector was clicked, filter will be emitted on a dashboard'));
     }
   };
+
+  const chartMargin = showLabels ? LABELS_MARGIN : 20;
 
   const handleLegendClick = ({ id }: EventData) => {
     let resultKeys = [];
@@ -137,26 +159,23 @@ const Pie: FC<PieProps<string, string>> = memo(props => {
   };
 
   const chartHeight = height;
-  const outerRadius = (width < chartHeight ? width : chartHeight) / 2 - (showLabels ? LABELS_MARGIN : 20);
+  const outerRadius = (width < chartHeight ? width : chartHeight) / 2 - chartMargin;
   const chartWidth =
-    isSideLegend && legendWidth
-      ? Math.max((outerRadius + (showLabels ? LABELS_MARGIN : 20)) * 2 + legendWidth, width)
-      : width;
+    isSideLegend && legendWidth ? Math.max((outerRadius + chartMargin) * 2 + legendWidth, width) : width;
 
   const pieProps: RechartsPieProps & { key?: string | number } = {
     activeIndex: activeIndex,
-    key: exploreCounter,
+    key: updater,
     data: currentData,
     dataKey: dataKey,
-    cx: isSideLegend ? outerRadius + (showLabels ? LABELS_MARGIN : 20) : '50%',
+    cx: isSideLegend ? outerRadius + chartMargin : '50%',
     outerRadius,
-    // @ts-ignore
     label: showLabels
-      ? // @ts-ignore
-        (labelProps: PieLabelRenderProps) => renderActiveShape({ ...labelProps, groupBy, pieLabelType })
+      ? labelProps => renderActiveShape({ ...labelProps, groupBy, pieLabelType } as ActiveShapeProps)
       : false,
     onClick,
   };
+
   if (isDonut) {
     pieProps.activeShape = activeShapeProps =>
       renderActiveShape({ ...activeShapeProps, groupBy, pieLabelType, isDonut: true });
@@ -168,32 +187,30 @@ const Pie: FC<PieProps<string, string>> = memo(props => {
   return (
     <Styles height={height} width={width} legendPosition={legendPosition} ref={rootRef}>
       {notification && <Notification onClick={closeNotification}>{notification}</Notification>}
-      <ResponsiveContainer width={chartWidth}>
-        <PieChart key={exploreCounter + chartWidth} width={chartWidth}>
-          {showLegend && (
-            <Legend
-              onClick={handleLegendClick}
-              {...getLegendProps(legendPosition, height)}
-              iconType="circle"
-              iconSize={10}
-              payload={data.map(item => ({
-                value: item[groupBy],
-                id: item[groupBy],
-                payload: item,
-                type: disabledDataKeys.includes(item[groupBy]) ? 'line' : 'circle',
-                color: CategoricalColorNamespace.getColor(item[groupBy], colorScheme),
-              }))}
-            />
-          )}
-          {((isSideLegend && legendWidth) || !isSideLegend) && (
-            <RechartsPie onAnimationEnd={() => setIsAnimationEnded(true)} {...pieProps}>
-              {currentData?.map((entry, index) => (
-                <Cell fill={getColor(entry[groupBy], colorScheme)} />
-              ))}
-            </RechartsPie>
-          )}
-        </PieChart>
-      </ResponsiveContainer>
+      <PieChart key={`${updater}`} width={chartWidth} height={height}>
+        {showLegend && (
+          <Legend
+            onClick={handleLegendClick}
+            {...getLegendProps(legendPosition, height)}
+            iconType="circle"
+            iconSize={10}
+            payload={data.map(item => ({
+              value: item[groupBy],
+              id: item[groupBy],
+              payload: item,
+              type: disabledDataKeys.includes(item[groupBy]) ? 'line' : 'circle',
+              color: CategoricalColorNamespace.getScale(colorScheme)(item[groupBy]),
+            }))}
+          />
+        )}
+        {((isSideLegend && legendWidth) || !isSideLegend) && (
+          <RechartsPie {...pieProps}>
+            {currentData?.map((entry, index) => (
+              <Cell fill={CategoricalColorNamespace.getScale(colorScheme)(entry[groupBy])} />
+            ))}
+          </RechartsPie>
+        )}
+      </PieChart>
     </Styles>
   );
 });