+++ plugins/plugin-chart-pie/test/plugin/transformProps.test.ts
@@ -1,4 +1,4 @@
-import { ChartProps } from '@superset-ui/chart';
+import { ChartProps } from '@superset-ui/core';
 import transformProps from '../../src/plugin/transformProps';
 
 xdescribe('Pie tranformProps', () => {