+++ plugins/plugin-chart-pivot-table/src/components/ColumnsHeader.tsx
@@ -1,4 +1,21 @@
-// @ts-ignore
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
 import React, { FC } from 'react';
 import { GridItem } from './Layout';
 import { Unit } from '../plugin/utils';
@@ -19,35 +36,20 @@ const ColumnsHeader: FC<ColumnsHeaderProps<string, string>> = ({
   <>
     {metrics.map(metric => (
       // eslint-disable-next-line react/jsx-key
-      <GridItem
-        bordered
-        header
-        bgLevel={2}
-        gridColumn={`span ${numberOfColumns / metrics.length}`}
-      >
+      <GridItem bordered header bgLevel={2} gridColumn={`span ${numberOfColumns / metrics.length}`}>
         {metric}
       </GridItem>
     ))}
     {columns.map(column =>
       uiColumnUnits[column].map(item => (
         // eslint-disable-next-line react/jsx-key
-        <GridItem
-          header
-          bordered
-          bgLevel={2}
-          gridColumn={`span ${numberOfColumns / uiColumnUnits[column].length}`}
-        >
+        <GridItem header bordered bgLevel={2} gridColumn={`span ${numberOfColumns / uiColumnUnits[column].length}`}>
           <div>{item}</div>
         </GridItem>
       )),
     )}
     {/* One empty line for header columns of rows */}
-    <GridItem
-      header
-      bordered
-      gridColumn={`span ${numberOfColumns}`}
-      style={{ color: 'transparent' }}
-    >
+    <GridItem header bordered gridColumn={`span ${numberOfColumns}`} style={{ color: 'transparent' }}>
       :)
     </GridItem>
   </>